Log errors when opacity not in DB.
I found a problem where errors were not logged when the
zenith opacity could not be found because the observation date
was too early.

I also cleaned up an error when tyring to remove the summary/
directory when it wasn't empty.

Finally, I updated a couple method names for the newer version
of fitsio.

+++ src/contrib/make_sdf_and_dbcon.py
@@ -334,7 +334,8 @@ def dbcon(sys):
         [os.unlink(xx) for xx in sdf_files]
         if os.path.isdir('summary'):
             [os.unlink('summary/'+xx) for xx in os.listdir('summary')]
-            os.rmdir('summary')
+            if not os.listdir('summary'): # rmdir if empty
+                os.rmdir('summary')
     
     
 if __name__ == '__main__':
